refactor(api): extract http.go into modular components
Split the monolithic http.go (1,007 lines) into focused modules following
separation of concerns:

- deps.go (53 lines): Interface definitions (Logger, Metrics, Auth, etc.)
- errors.go (59 lines): Response helper functions (writeJSON, writeError, etc.)
- auth.go (77 lines): Authentication middleware with constant-time comparison
- fileserver.go (242 lines): Secure file serving with path traversal protection
- http.go (reduced from 827 to 600 lines, -27%): Core routing and handlers

Security improvements:
- Multi-layer path traversal protection (URL decode, Unicode normalization)
- Symlink resolution and containment verification
- ETag-based HTTP caching for file serving
- Constant-time token comparison to prevent timing attacks

Test coverage maintained:
- All existing API tests pass (api, api/v1, api/middleware)
- Added SetStatus/SetRefreshFunc test helpers
- Verified backward compatibility with legacy endpoints
